requirements.txt
paho-mqtt~=1.6.1
numpy~=1.24.1
Pillow~=9.4.0
picamera~=1.13